Change request header from Authorization to X-API-Key
From testing the endpoint, the platform API endpoint called returns a 403 (Forbidden)

We think that the request header being generated is of the format:
Key: Authorization, Value: X-API-Key some_secret_string which the platform API doesn't accept.

This change should create a request header of the format:
Key: X-API-Key, Value: some_secret_string which is what the platform API is expecting.