Remove leading '/' from path created when getting a case note by id
When testing the service API endpoint to get a specific historic case note by
its id, the platform API endpoint called returns a 403 (Forbidden).

From inspecting the code, we found that the endpoint that the service API
was trying to call was different to the actual platform API endpoint for getting a
specific case note.

We noticed that the path created for getting all cases notes did not have a
leading '/' and works when called.