Add feature flag condition to the List Case Notes By Person Id endpoint
This adds a condition in the controller action for the list case notes by person Id that checks the value of the feature flag.

If the feature flag is set to true, the usecase is called and the expected functions are triggered.
If the feature flag is set to false, null or anything else, the response returns a 500 status code.