Check feature flag value before calling platform gateway in Process Data Gateway
The process data gateway also attempts to make calls to the social care platform gateway so that it can return historic case notes or visits as part of the records history.

This adds a condition within the process data gateway to check the value of the feature flag before attempting to retrive historic mosaic data via the platform gateway.