Update the feature flag for production to be created set as true
We created secrets for both staging and production that would store values for our feature flag functionality.

Initially, set the value of the flag in production to be 'false', just in case any production releases needed to happen while we were still testing feature in staging.

There hasn't been any production releases yet and the feature has been approved to go live, so updating the value of the secret so that it's created set as 'true' instead.