Update HistoricalCaseNotesToDomain to return note type for form type
In the records history for a person, the “Record type” or `form_name`
or “form type” returned for historic case notes is set to the `title`
of it. This makes it difficult to filter by type on the frontend
because the type for a historic case note is not generic.