Add feature flag for using note type for form name in historic case notes
In a previous PR #202, we aimed to fix the fact that historic case notes were
returning the incorrect data for `form_name` i.e. record type by using the
case note title instead of note type for the `api/v1/cases` endpoint i.e.
records history in the frontend. This meant that it would be aligned with
other records and with the `api/v1/casenotes/{id}` endpoint which is correctly
returning the note type for form name.

However, it's been raised that this should be feature flagged for now
as we don't want to yet to make this change.