Use random string with max length for generated ethnicity property
It looks like a couple of gateway tests fail intermittently.

Error generated: Npgsql.PostgresException : 22001: value too long for type character varying(33)

This could be as a result of the faker class generating an ethnicity that is greater than the max length
defined in the schema for that column.

This replaces the faker method `country` with a random string with a max length value for the generated ethnicity,
so that when this property is saved in the database, the value saved would be less than the max length.