Edit functionality to return the warning note with any saved reviews
Initially, the gateway was going to be set up
to return just the reviews saved for a specific warning note.

After reviewing (no pun intended!)
the usecase where we would need to return these warning note reviews,
we decided that the gateway (and overall functionality of the endpoint)
should return a specific warning note
and any saved reviews created for that warning note.

This adds a "list of reviews" attribute to the WarningNote domain class.

The gateway action when given an ID
will return a response mapped to the WarningNote domain object
(or null if no matches are found)