Add usecase & controller to retrieve a warning note
Adds a usecase that throws an exception if no warning note is returned from the gateway.

The usecase also maps the domain object returned from the gateway to a response object.

The response object is similar in structure to the domain object with the exception being that all
date related attributes are converted to strings and follow the ISO 8601 format.

The response (or thrown exception) from the usecase would be returned to the controller.