Add missing Review Date & Created By properties
When get make a request to retrieve warning notes for a specific person,
we map the matches returned from the database to a domain object.

The domain object was missing the `Review Date` & `Created By`
properties, which are variables that can given as part of the POST request
i.e when the warning note is initially created.