Remove leading '/' from the route
Currently troubleshooting why the endpoint returned a 404 when called and one of the errors
in the lambda logs states: "Request does not contain domain name information but is derived from APIGatewayProxyFunction."

From inspecting the code found that there was a leading '/' in the route name and for the other routes we don't have a leading '/'.

This also changes the "warningNotes" to "warningnotes" in the path.