Save the decision to disclose a warning note review
Previously, we were not retrieving nor saving a value for `DisclosedWithIndividual`
on a warning note review.

In the frontend, a form element was added that would ask a user reviewing
an ADULT warning note only, if they had disclosed the review to the individual.

When this form is submitted, the frontend would be making a
PATCH request that would include a `DisclosedWithIndividual` value.

This change updates the service API to recieve this updated request and process it.