Fix not getting related entities to personal relationships
We made an assumption that by just including `PersonalRelationships`,
it would also include the related entities within a personal relationship
like the `Type` and `OtherPerson` which seemed like right as our tests
passed.

However, when testing the API endpoint we realised this was incorrect
and our tests only passed because EF Core was tracking the navigation
properties because we had just seeded those entities. Therefore, by
disabling tracking, our tests fail as expected.