Change EndedDate validation
This would check that the datetime sent is not from the next day onwards
so would allow a datetime set for today even if the datetime from the
client is slightly greater than the server's

This is an attempt to fix a bug where ending a warning note fails because
the ended date passed in by the client is not less than the date time of when the
code in the backend validator executes.

This change should allow any attempt to end a warning note as long as the datetime
is today or in the past.

We currently do not have functionality to allow a user to add a custom warning note date,
so the assumption is when they select end a warning note, it is an immediate desire.