chore: update the make restart db command
The make command to restart the test databases was referencing
the old image name of the postgres test database.

This updates the command to reference the new image and service name for the docker test database.

This might not work if the old `test-database` image still exists
or if the new image name: `sccv-api-test-postgresql` does not exist.

Would recommend checking and cleaning up the docker images to match what we be expecting before running this command.